Add Balance and Lib_Controller
Adding my first pass at a pid-controlled auto-righting script. It keeps
the craft pointed UP regardless of input. Ideally this would also arrest
horizontal motion but that's later.

Also added lib_controller, which sets up axis mappings and a way to
display those axis mappings for AGs 1 to 8, allowing for switches. Would
also like to set up this or something similar to be action-mapped so
that a single button press does a thing once. Future iterations.