Hover updates, New Lib
Also fixed controller so the visual state of controls can be driven on
an as-needed basis, instead of forcing the idea of toggle switches. Now
control visualization can be a mix of toggle and push.

Set the landAnywhere function to switch to hover mode at cutoff height
to better control landing.

Added in hooks for the hover script to have a horizontal velocity kill
mode. Logic behind that is still definitely in progress.